requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
freeze==3.0
idna==2.10
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
